refactor(utils/customer): move meal evaluation and easter egg check logic to instances
